# backend/src/design/api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: Learnify API
  version: "3.0"
  description: |
    Online Course Platform API — v3.
    Changes from v2:
    - Removed dead endpoints: /courses/featured, /courses/continue, /courses/stats, /lessons/:id/course, /payments/webhook-logs
    - Password validation: min 8 chars, 1 uppercase, 1 number
    - Email uniqueness check on profile update (409 if taken)
    - Pagination params clamped (page >= 1, limit 1-100)
    - Notifications response now includes pagination object
    - Course list: added enrollment, price, instructor_id filters
    - Webhook: improved auth validation, atomic payment update
    - Comments: lesson existence check, batched mention queries, content type validation
    - Bookmarks: course existence check (404)
    - Enrollments: progress object validation
    - Course delete: cascade deletes comments, mentions, notes, reviews, bookmarks
    - Upload: file type validation (MIME whitelist)
servers:
  - url: https://asm3-cos30043.vercel.app
    description: Production (Vercel)
  - url: http://localhost:3000
    description: Local dev
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [student, instructor] }
        bio: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
    Comment:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        lesson_id: { type: string, format: uuid }
        parent_id: { type: string, format: uuid, nullable: true }
        content: { type: string }
        username: { type: string }
        avatar_url: { type: string, nullable: true }
        reply_count: { type: integer }
        replies:
          type: array
          items: { $ref: "#/components/schemas/Comment" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Pagination:
      type: object
      description: "page >= 1, limit clamped to 1-100"
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer }
        pages: { type: integer }
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8, description: "Min 8 chars, 1 uppercase, 1 number" }
                role: { type: string, enum: [student, instructor], default: student }
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      token: { type: string }
                      user: { $ref: "#/components/schemas/User" }
        "400": { description: Missing/invalid fields }
        "409": { description: Email already registered }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      token: { type: string }
                      user: { $ref: "#/components/schemas/User" }
        "400": { description: Missing fields }
        "401": { description: Invalid credentials }
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/User" }
  /api/auth/search:
    get:
      tags: [Auth]
      summary: Search users (for @mentions)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Username prefix search
        - name: course_id
          in: query
          schema: { type: string, format: uuid }
          description: Limit to users enrolled in or instructing this course
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        username: { type: string }
                        avatar_url: { type: string, nullable: true }
                        role: { type: string }
...
  /api/lessons/{lessonId}/comments:
    get:
      tags: [Comments]
      summary: List top-level comments with replies (threaded)
      parameters:
        - name: lessonId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Threaded comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      comments:
                        type: array
                        items: { $ref: "#/components/schemas/Comment" }
                      pagination: { $ref: "#/components/schemas/Pagination" }

    post:
      tags: [Comments]
      summary: Post comment (supports @mentions and replies)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, description: "Use @[username] for mentions" }
                parent_id: { type: string, format: uuid, description: "Reply to this comment" }
      responses:
        "201": { description: Comment created (notifications sent for mentions/replies) }
        "400": { description: Content required }
        "404": { description: Lesson not found }
  /api/comments/{id}:
    put:
      tags: [Comments]
      summary: Edit own comment
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
      responses:
        "200": { description: Updated }
        "403": { description: Not owner }
        "404": { description: Not found }

    delete:
      tags: [Comments]
      summary: Delete own comment
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Deleted }
        "403": { description: Not owner }
        "404": { description: Not found }

  # ── Notes ────────────────────────────────────────────────
...